requirements.txt
Package                   Version
------------------------- --------------
fastapi                   0.111.0
paho-mqtt                 2.1.0
pika                      1.3.2
pymongo                   4.8.0
uvicorn                   0.30.1